Initial commit
Initial commit with placeholders for all logic functioning.

The warning counter program searches for files, parses some configs
containing regexes, match these regexes in the found log files, compares
the number of found warnings against the defined thresholds in found
limit files.

It's mostly ad-hoc, unstable and untested, but the bulk of the logic is
in place.