Fix issue with not converting to wildcard categories early enough
Because the threshold comparisons were done on LimitEntries gotten from
the parsed logfiles, and not on the ones gotten from the parsing of the
limits files, we would incorrectly, compare the number of warnings
against the threshold individually on multiple categories, when they
should in fact have been merged into one set.

Assume we found these warnings
/tmp/Limits.toml[gcc/cat1]:5:5
/tmp/Limits.toml[gcc/cat1]:5:6
/tmp/Limits.toml[gcc/cat2]:5:5

But /tmp/Limits.toml were in face defined like so:
gcc = 2,
Then, individually, neither cat1 nor cat2 would breach the limits, but
combined, they do.

This change fixes that issue by post-processing the gathered results,
and ensuring they are keyed by the entries that exists in the flattened
limits.