Add option to prune Limits files
When a per-category limit reaches zero, we can remove the category
listing entirely. When there is only the one category left (or all are
zero), we can reduce the per-category listing into a simple listing.

This change also removes the `default` feature, as its interaction with
this pruning caused problems. The default feature also goes against what
I want this tool to be used for.