Replace ignore with jwalk
I realized that we don't want the .gitignore features of ignore, so
we might as well use jwalk. (Because build dirs are usually ignored..)

I've also realized a major design flaw where the functions that return
Receiver<T>s doesn't actually process stuff in the background, meaning
we risk deadlocking the program if the channels run out of capacity. We
need to streamline the streaming somehow..