Fix threading issue in search_in_files
The design using rayon::scope had a fatal flaw where it would produce
all values before the function returned. This of course is bad, as we
want to start consuming as quickly as possible.

This change fixes that by using std::thread::spawn and start a true
background thread. It's just sad that we cannot share data as
efficiently as before.